Refactored store to use state & fix nav layout shift on load
