+++ src/lib/components/Toast_v2.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import CloseButton from './Buttons/CloseButton.svelte';
 	import { fade } from 'svelte/transition';
-	import { restTimer$, type Toast } from '$lib/store';
+	import { restTimer, type Toast } from '$lib/state.svelte';
 
 	interface Props {
 		type?: Toast['type'];
@@ -14,11 +14,11 @@
 </script>
 
 {#if type === 'rest'}
-	{#if $restTimer$}
+	{#if restTimer.value}
 		<div
 			class="alert alert-info flex min-w-36 justify-center gap-2 bg-neutral text-neutral-content"
 		>
-			<span>{$restTimer$}</span>
+			<span>{restTimer.value}</span>
 			<small> Next set</small>
 		</div>
 	{/if}
