+++ src/lib/components/Toasts.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
-	import { dismissToast, toasts$ } from '$lib/store';
+	import { toaster } from '$lib/state.svelte';
 	import Toast from './Toast_v2.svelte';
 </script>
 
 <div class="toast">
-	{#each $toasts$ as toast (toast.id)}
+	{#each toaster.getToasts() as toast (toast.id)}
 		<Toast
 			type={toast.type}
 			dismissible={toast.dismissible}
 			message={toast.message}
-			dismiss={() => dismissToast(toast.id)}
+			dismiss={() => toaster.dismissToast(toast.id)}
 		/>
 	{/each}
 </div>
